Rename the API models package
