+++ src/api/model/external_tools_response.py
