+++ test/api/mapper/test_user_mapper.py
@@ -3,7 +3,7 @@ from datetime import date
 from uuid import UUID
 
 from api.mapper.user_mapper import api_to_domain, domain_to_api
-from api.models.user_settings_payload import UserSettingsPayload
+from api.model.user_settings_payload import UserSettingsPayload
 from db.model.user import UserDB
 from db.schema.user import User
 from util.functions import mask_secret
