+++ tools/db_apply_head_schema.sh
@@ -9,7 +9,7 @@ if [ ! -f "Pipfile" ]; then
 fi
 
 if [ "$1" != "-y" ]; then
-    echoinfo "Have you created all previous schemas and applied them to the database? (Y/n) "
+    echoinfo "Have you created all previous schemas and applied them to the database? (y/n) "
     read -r RESPONSE
     if [ ! "$RESPONSE" = "Y" ]; then
         echowarn "Let's make sure that's done first."
